backend/requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==4.2.7
django-cors-headers==4.2.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
et_xmlfile==2.0.0
gunicorn==23.0.0
numpy==2.2.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
Pillow==10.1.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
setuptools==78.1.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
waitress==3.0.2
XlsxWriter==3.2.2
